move example out, remove duplicated plot utils
