+++ examples/acados_python/chain_mass/solution_sensitivity_example.py
